requirements.txt
# requirements file
# Python version: 3.10.13
# Dependencies for the sequential-chat-agents project

pyautogen==0.2.25
google-generativeai==0.8.3
python-dotenv==1.0.1
pyyaml==6.0.1
backoff==2.2.1
python-dotenv==1.0.1
